lodie_extract/requirements.txt
dateutils
BeautifulSoup4
lxml
tldextract
IPython==1.2.0